allow name in component_from_yaml
