+++ pp/components/test_properties/test_components_ports_mzi_arm_.csv
@@ -1,7 +1,7 @@
 ,W0,E0,E_0,E_1,E_2,E_3
-0,6.1232339957367663e-16,49.999999999999986,40,10.000000000000004,10,39.999999999999986
-1,0,-1.7763568394002505e-15,73.5,73.5,6.5,6.5
-2,180,0,90,90,270,270
+0,-49.999999999999986,3.5527136788005009e-15,-9.9999999999999893,-39.999999999999986,-39.999999999999986,-9.9999999999999964
+1,80,80,6.5,6.5,73.5,73.5
+2,180,0,270,270,90,90
 3,0.5,0.5,10,10,10,10
 4,1,1,49,49,49,49
 5,0,0,0,0,0,0
