+++ pytest.ini
@@ -1,14 +1,15 @@
 [pytest]
-testpaths = pp
+testpaths = pp/
 # addopts = --tb=no
 addopts = '--tb=short'
 python_files = pp/*.py notebooks/*.ipynb
-nb_file_fnmatch = notebooks/*.ipynb
-nb_test_files = True
-nb_exec_allow_errors = True
-nb_exec_timeout = 100
-nb_post_processors = blacken_code
-nb_coverage = True
+
+; nb_file_fnmatch = notebooks/*.ipynb
+; nb_test_files = True
+; nb_exec_allow_errors = True
+; nb_exec_timeout = 100
+; nb_post_processors = blacken_code
+; nb_coverage = True
 
 
 filterwarnings =
